airavata-python-sdk/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "airavata-python-sdk"
version = "3.0.0"
description = "Apache Airavata Python SDK"
readme = "README.md"
license = "Apache-2.0"
authors = [{ name = "Airavata Developers", email = "dev@airavata.apache.org" }]
requires-python = ">=3.10"
dependencies = [
  "oauthlib",
  "requests",
  "requests-oauthlib",
  "grpcio>=1.60.0",
  "protobuf>=4.25.0",
  "googleapis-common-protos>=1.62.0",
  "paramiko",
  "scp",
  "pysftp",
  "configparser",
  "urllib3",
  "pyjwt",
  "pydantic",
  "rich",
  "pandas",
  "pyyaml",
]

[project.optional-dependencies]
notebook = [
  "ipywidgets",
  "ipython"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["airavata_sdk*"]
exclude = ["*.egg-info"]

[tool.setuptools.package-data]
"airavata_sdk.transport" = ["*.ini"]
"airavata_sdk.samples.resources" = ["*.pem"]
